reactor: convert task quota signal to a timerfd+thread
In Linux, signals are protected by a global lock, and thus don't scale with
increasing core counts.  We already observed this on large machines, and
it will grow worse with our plans for sub-task-quota timing with the
CPU scheduler.

To prevent a lock meltdown, convert the signal to a timerfd + a thread
to wait for it and update g_need_preempt.  We pay the cost of a context
switch, but that is not very different from signal delivery, (perhaps even
cheaper), and scales with increasing core counts.

In the future we may want to have timer threads service g_need_preempt
for other reactors on the same processor socket, if it proves necessary.

(cherry picked from commit f03ebd38d5108a3709d76cd12065ba34fcba2bbe)
Signed-off-by: Glauber Costa <glauber@scylladb.com>