Prometheus: do not rely on collectd for the hostname
If collectd is not started the hostname in collectd will be set to the
default 'localhost', this will cause all the servers to report as
localhost instead of their true name.

This patch adds a hostname to the prometheus config, if the hostname is
not initilized the system hostname is taken.

Fixes #233

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <1483886036-27692-1-git-send-email-amnon@scylladb.com>