metrics: Make the instance_id a label
This patch is the first step to remove the instance_id completely.
The instance_id (a heritage from collectd) marks the shard a metrics was
generated from.

Now that the metrics layer support labels, it should be store as a
label, the collectd will treat it differently in its own internal
implementation.

This patch contains multiple chagnes:
* The labels are represented better as a sorted map, as there can be
  only one label per type, they should always be sorted and it's useful
  to look for a specific label.
* The metrics layer will check in the labels if a shard label was
  define, if not it will take the current shard (the same as the current
  implementation).
* The collectd and prometheus implementation were modified to support
  this chagne.

An additional clean up will be made to remove the instance_id completely
from the API when all the code will migrate to support it.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>