http: break dependencies of the connection and the reply
The reply will need the connection object and the http would need the
reply.

To prevent header loop, this patch moves the connection object outside
of the http object, so it will be possible to forward declare it.

It also replaces the include of the reply with a forward decleration and
moving the functionality to the source file.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>