tests/http: Add a test for chunked body transfer
This test uses a client server helper testing function that works on the
connection level.

The test function creates two threads one for the server and one for the
client, and accept two lambda function.
The server function returns a vector of replies.
The client function gets each of the reply and an id of the current
message (so it can be stateless) each call the client return true if it
wants to continue or false if it completed, the client function does the
validation that the message is as expected.

Three tests were added:

1. a simple chunked test: a 100,000 long message is sent from the client
 to the server.

2. an error during body written, the body writer, first write 3000 bytes
 then throw an exception.

3. multiple messages are written with multiple sizes all success

To allow adding a listener to the http by the test, a test friend
class is used.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>