posix_net: load-balance connections
Currently we employ a quite naive load balancing, keep track of the
current shard receiving a connection with a global variable and
distributing new connections in a round-robin fashion.

However, connections coming from different server_socket's are very
different from each other in nature: some, like the prometheus metrics
connections tend to be short lived and low bandwidth, while others like
the application connections could be long lived and drive very high
levels of activity.

As a result of that, heavy traffic connections can end up bundled in a
small group of shards, which will have disproportionally higher load.

The solution for that is to move load balance as part of the instance,
instead of using a global variable. This way, things that listen()
together load balance together().

Also, connections can come and go. Simply round-robin'ing them as they
are accept'd do not account for that. We'll then keep track of the
connected_sockets being generated and destroyed, and pick the next CPU
based on that.

The result is a much fairer CPU utilization among shards.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20170519070112.13010-1-glauber@scylladb.com>
(cherry picked from commit 1e51fb2d49b8bc25a02de009246afa38068814e9)