Prometheus: avoid merging all metrics family
The prometheus server sends information from all shards, to avoid
long stalls of the shards, it uses continuation to send the response.

This patch change the way the metric_family_iterator works.

Instead of combining all the metrics from all shards at the beginning,
now, each iteration the iterator and metric_family create a combine
metrics.

So the work for each of the metrics family is done in a separate
continuation.

As part of this change the the data is split between
metric_family_iterator that holds the state and metric_family that act
as a facade class to a metric family.

Fixes #327

Message-Id: <20170827150130.12239-1-amnon@scylladb.com>