Remove disabled smp metrics
The metrics layer is now non tollerant to double registration,
The smp queue metrics are only registered but never disable, still they
cause an exception because they are using the same name, that is no
longer unique.

In Master, we use label to differ the queue, but here they are disable
anyhow and cannot be used, so it's cleaner to just remove them.

See Scylladb/scylla-enterprise#441

Message-Id: <20171224103222.16858-1-amnon@scylladb.com>