Merge NOWAIT AIO from Avi
Starting with Linux 4.14, a facility for not blocking during io_submit() is
included.  By setting the RWF_NOWAIT flag, the caller requests an error return
instead of blocking. This is very beneficial to seastar as blocking in the
reactor also blocks unrelated operations, killing concurrency.

Because libaio does not support this flag (and since its maintainer is not
responsive), libaio is replaced by a seastar wrapper. The wrapper is also
able to bypass the io_getevents() system call when a zero timeout is
specified (which is always, in seastar, so far).

The RWF_NOWAIT flag is detected during both compile-time and run-time.

* https://github.com/avikivity/seastar nowait-aio/v3:
  core: add alternative interface to Linux aio
  linux-aio: try to perform io_getevents in userspace
  Switch Linux-aio implementation away from libaio
  linux-aio: add support for NOWAIT aio
  reactor: provide stable iocb addresses
  reactor: support for NOWAIT aio

(cherry picked from commit 6972a1e839d0e2ec34acacd600bd6a76c3b9066c)