fstream: avoid false positives from Scylla sstable integrity detector
Scylla's sstable integrity detector warns if a file containing a 4k
block of zeros is found, before or after the write. However this can
happen in the following scenario:

 1. write a file of size 1 (mod 4096)
 2. the last byte happens to be zero
 3. fstream pads the last buffer to be 4096 byte size aligned
 4. the padding happens to be all zeros

If the above happens, the integrity checker (which is implemented as
the file the fstream writes to) will see a 4k pre-write block, and warn.

We observed such a false positive from a customer.

To silence the false positive, ensure the padding is not all zeros.

A better fix is likely needed for upstream, but this band-aid will
help prevent false positive at the customer.
Message-Id: <20180109113350.28970-1-avi@scylladb.com>